requirements.txt
# Core ML
pandas
numpy
scikit-learn
xgboost
lightgbm
pyarrow

# MLflow
mlflow

# API
fastapi
uvicorn[standard]
pydantic

# Database
sqlalchemy
psycopg2-binary
pandas-gbq
google-cloud-bigquery

# Data validation & drift
scipy

# Monitoring
prometheus-client

# Scheduling / automation
prefect>=2.14.0
prefect-sqlalchemy

# Data & utilities
python-dotenv
joblib
requests

# Testing
pytest
httpx
pytest-timeout

# Notebook
jupyter
matplotlib
seaborn
plotly